Fix the error logging in main.c
We can't use LOG() there, because the logging mechanism is not
initialized yet.

Also, fix the parameter handling. Before this commit, if the root
dir was passed before the parameters, they were ignored. Current version
is able to detect this situation and return an error message.

On Linux, getopt automatically puts non-parameter arguments to the end
of argv, so mixing parameters and root dir is still allowed.